move user data script logic to template
